preparing Room DDBB

+++ app/src/main/java/com/jorgesm/compose_marvel_api/utils/Routes.kt
@@ -3,6 +3,6 @@ package com.jorgesm.compose_marvel_api.utils
 sealed class Routes(val route: String) {
     object CharactersListScreen:Routes("charactersList")
     object DetailScreen: Routes("detail/{itemId}"){
-        fun addItemId(itemId: Int)= "detail/$itemId"
+        fun addItemId(itemId: Long)= "detail/$itemId"
     }
 }
